Crate cipha_lib

This is the main module for the cipha-lib crate. It contains various cipher and crypto functions.

Modules

ciphers

Functions

alpha2num

Converts alphabetic characters in the input text to their corresponding numerical values.

atbash_cipher

Function to encipher a given plaintext using the Atbash cipher.

atbash_decipher

Function to decipher a given ciphertext using the Atbash cipher.

caesar_cipher

Applies the Caesar cipher to the input message with the given shift.

create_morse_code_map

create_reverse_morse_code_map

morse_code_cipher

Encodes the input message into Morse code.

morse_code_decipher

num2alpha

Converts numerical values in the input text back to their corresponding alphabetic characters.

rail_fence_cipher

Function to encipher a given plaintext using the Rail Fence cipher.

rail_fence_decipher

Function to decipher a given ciphertext using the Rail Fence cipher.

reverse_cipher

Reverses the input message.

rot13

Applies the ROT13 cipher to the input message.

vigenere_cipher

Encrypts the input message using the Vigenère cipher with the given key.

vigenere_decipher